% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{append_clusters}
\alias{append_clusters}
\title{Clusters the t-SNE embedding of EMERGe data.}
\usage{
append_clusters(X, minPts = 10)
}
\arguments{
\item{X}{(tibble) EMERGe data with at minimum 'n10', 'tsne1', and 'tsne2'
columns.}

\item{minPts}{(integer) Minimum number of points necessary to define a
cluster.}
}
\value{
Provided X (tibble) with cluster assignment appended as a
'cluster' column.
}
\description{
Runs Hierarchical Density-Based Clustering with Applications to Noise
(HDBSCAN) to cluster the t-SNE embedding of EMERGe data.
}
