% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca.R
\name{append_mca}
\alias{append_mca}
\title{Transforms cleaned EMERGe data into a richer representation using
multiple correspondence analysis (MCA).}
\usage{
append_mca(X, nf = 50, top_quantile = 0.999, get_eigs = FALSE)
}
\arguments{
\item{X}{(tibble) EMERGe dataset with 'n10' and 'postmean' columns.
The 'postmean' column may be obtained from emergeR::stats().}

\item{nf}{(integer) default 50, maximum number of factors kept from MCA
(dudi.acm)}

\item{top_quantile}{(float) default 0.999, defines the percentile by which
sequences are considered to be top performers.
For a dataset with ~900,000 n10 sequences, the default
value selects ~900 top sequences for further analysis.}

\item{get_eigs}{(logical) default FALSE, return eigs in a list with the
MCA tibble for debug purposes.}
}
\value{
A list with two components:
\describe{
\item{edit_top}{A tibble containing the subset of EMERGe data
corresponding to sequences in the top quantile, joined with
their MCA coordinates.}
\item{eigens}{A numeric vector of eigenvalues from the MCA,
representing the variance explained by each axis.}
}
}
\description{
Similar to PCA, MCA identifies the main axes of variation across
categorical variables and projects the data along them.
}
