% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{append_stats}
\alias{append_stats}
\title{Calculate statistics on a cleaned EMERGe dataset.}
\usage{
append_stats(X_clean, num_folds = 5, seed = 1, get_cv_stats = FALSE)
}
\arguments{
\item{X_clean}{(tibble) Cleaned data from emergeR::clean.R.}

\item{num_folds}{(integer) Default 5, number of folds to use for k-fold CV.}

\item{seed}{(float) Default 1, controls RNG for reproducibility.}

\item{get_cv_stats}{(logical) Default FALSE, returns CV diagnostics if TRUE.}
}
\value{
If get_cv_stats is TRUE, a list with two components:
\describe{
\item{X_stats}{A tibble containing EMERGe data: n10, mle, upper_ci,
lower_ci, postmean, upper_cred, lower_cred, alpha, beta.}
\item{cv_stats}{A tibble containing alpha MLE, beta MLE, number of
samples in the held-in fold (test_n), log-likelihood of the
test data (logLik), mean negative log-likelihood per observation
(meanNegLL), Akaike Information Criterion for the fit (p = 2)
(AIC), Pearson chi-square statistic comparing observed vs.
expected counts (PearsonChi2), Root mean squared error between
observed and expected counts (RMSE_counts), Mean absolute
error between observed and expected counts (MAE_counts),
Optimizer convergence code (0 indicates success) (convergence),
and the Optimizer convergence message (message).}
}
Otherwise, only X_stats is returned.
}
\description{
Calculates editing MLE, Wilson CIs (binomial), posterior mean under an
empirical Bayes Beta prior (with k-fold CV), and Bayesian credible intervals.
}
\examples{
\dontrun{
emergeR::stats(X_clean = clean_data)
emergeR::stats(X_clean = clean_data, num_folds = 4)
emergeR::stats(X_clean = clean_data, num_folds = 4, seed = 42)
}

}
