% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{append_tsne}
\alias{append_tsne}
\title{Dimensionally reduces MCA embedding of EMERGe data using t-distributed
stochastic neighbor embedding (t-SNE).}
\usage{
append_tsne(
  X_mca,
  alpha = NULL,
  h = NULL,
  seed = NULL,
  perplexity = 30,
  verbose = FALSE,
  max_iter = 30
)
}
\arguments{
\item{X_mca}{(tibble) EMERGe data with at least 'n10' and MCA columns that
start with "CS".}

\item{alpha}{(float) Exaggeration factor, handled internally by default.}

\item{h}{(float) Learning rate, handled internally by default.}

\item{seed}{(float) RNG seed, handled internally by default.}

\item{perplexity}{(float) effective number of nearest neighbors,
handled internally by default.}

\item{verbose}{(logical) whether to print Rtsne progress messages.}

\item{max_iter}{(integer) maximum number of iterations for
t-SNE optimization.}
}
\value{
X (tibble) with 2D t-SNE coordinates attached as 'tsne1'
and 'tsne2' columns.
}
\description{
Motivated by insufficient separability of clusters on EMERGe MCA data alone.
Calculates hyperparameters that work on EMERGe data if left NULL.
Calculates a t-SNE fit and appends t-SNE coordinates to each n10 in X.
}
