% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean an EMERGe dataset after HT Stream preprocessing.}
\usage{
clean(fastq_infile, template, flanks)
}
\arguments{
\item{fastq_infile}{(String) Path to a FASTQ file.}

\item{template}{Character(1) DNA template sequence.}

\item{flanks}{Character(4) flank subsequences in expected order,
(f1, f2, f3, f4).}
}
\value{
A tibble with N10-level summaries: number of reads "n,"
number of edits "k," and other codons "other."
}
\description{
Reads a FASTQ file and returns a standardized tibble.
}
\examples{
\dontrun{
# R255XE488Q. Target is GAA edited to GGA. We take 6-mer flanks around GAA 
# and the N10. The full RNA hairpin follows:
# GGGAATGGCGCTTCCCGGCA GGAAGT GAA AAGCTG AGGCCGACTTTCTTTCTTTCGTCG GCCTCA 
# NNNNNNNNNN TCCTGC CGGGTTGTAGGGTC
# Flanks, demarcated above by spaces: GGAAGT AAGCTG GCCTCA TCCTGC
template = paste0(
  "GGGAATGGCGCTTCCCGGCAGGAAGTGAAAAGCTGAGGCCGACTTTCTTTCTTTCGTCGGCCTCA",
  "NNNNNNNNNNTCCTGCCGGGTTGTAGGGTC"
)
flanks = c("GGAAGT", "AAGCTG", "GCCTCA", "TCCTGC")
emergeR::clean(
  fastq_infile = "example.fastq", template = template, flanks = flanks
)
}

}
