% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{plot_motif_logos}
\alias{plot_motif_logos}
\title{Plots a sequence logo per each identified EMERGe cluster.}
\usage{
plot_motif_logos(X_clustered)
}
\arguments{
\item{X_clustered}{(tibble) EMERGe data with at minimum 'n10' and 'cluster'
columns.}
}
\value{
(list of plots) A list of seq logos, sized according to the number
of clusters.
}
\description{
Plots a sequence logo per each identified EMERGe cluster.
}
