% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{plot_tsne_clusters}
\alias{plot_tsne_clusters}
\title{Plots t-SNE coordinates colored by cluster assignment.}
\usage{
plot_tsne_clusters(X)
}
\arguments{
\item{X}{(tibble) EMERGe data with at minimum 'tsne1', 'tsne2', and'cluster' columns.}
}
\value{
(plot) ggprism-formatted plot.
}
\description{
Plots t-SNE coordinates colored by cluster assignment.
}
