% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{plot_tsne_edits}
\alias{plot_tsne_edits}
\title{Plots t-SNE coordinates colored by posterior mean editing.}
\usage{
plot_tsne_edits(X)
}
\arguments{
\item{X_tsne}{(tibble) EMERGe data with at minimum 'tsne1', 'tsne2', and'postmean' columns.}
}
\value{
(plot) ggprism-formatted plot.
}
\description{
Plots t-SNE coordinates colored by posterior mean editing.
}
